refactor: CROP_LIST
- 아이콘 추가 및 기존 mockData받는 컴포넌트 CROP_LIST로 변경
